remove useless functions
Replace getdevice with UserInputService:GetPlatform.
Replace getping with `game:GetService("Stats"):FindFirstChild("PerformanceStats").Ping:GetValue()`, note that the value is a string instead of a number, so you'll need to use tonumber, and the ping is in milliseconds.
Replace getfps with the code from  https://devforum.roblox.com/t/get-client-fps-trough-a-script/282631